requirements.txt
streamlit
openai
pandas
matplotlib
seaborn
python-dotenv
notebook